Add files via upload
This has the "infrastructure" of the NN and uproot connected. BUT! at the moment, it is the kinematic data going into the net, which is not (only) what we need. Also, it would be nice to put vectors containing all the mentioned variables in the README of the ROOT data into the LDA classifier and see how that performs.